requirements.txt
Django==4.2.29
djangorestframework
djangorestframework-simplejwt
django-cors-headers
cloudinary
django-cloudinary-storage
dj-database-url
python-dotenv
psycopg2-binary
celery
redis
django-celery-results
adrf
channels
channels-redis
daphne
